Update Makefile to use .flake-proprietary config
The main st2 repo is refactored to use st2flake8 for license and copyright checks. There is a new .flake8-proprietary config file to check related repo for the proprietary license header. The Makefile for this repo is updated accordingly. The tests directory is also added in the flake8 check.